Split PlayerMovement into helper modules
