+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementCollisions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf533be1eb89c8a4f99afee60ac57d01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
