+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementDash.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e15cefdda316b53468eb4c9522198350
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
