+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementGravity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e208b78c961025842aed5ecfa7793889
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
