+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementHelpers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edc35c7d9f8f6fe4f8e27cd6f112aeab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
