+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f49fdbe78ac189647a303770f14c02da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
