+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 75ca8fe30bf39b14ea7f390156e1f635
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
