+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementRun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1cf122cbdc52be24fba86f5bfbec2b94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
