+++ Platformer Demo - Unity Project/Assets/Scripts/PlayerMovementSlide.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7af99151db00ab74582d166c9538fc16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
